🔥 重複して存在していたfrontendディレクトリを削除した
