+++ frontend/.prettierrc.json
@@ -4,5 +4,5 @@
   "semi": false,
   "singleQuote": true,
   "jsxSingleQuote": true,
-  "printWidth": 100 
+  "printWidth": 100
 }
