--- frontend_second/.eslintrc.json
@@ -1,20 +1,0 @@
-{
-  "extends": [
-    "next",
-    "next/core-web-vitals",
-    "prettier",
-    "next/babel",
-    "plugin:import/recommended",
-    "plugin:import/warnings"
-  ],
-  "rules": {
-    "import/order": [
-      "error",
-      {
-        "alphabetize": {
-          "order": "asc"
-        }
-      }
-    ]
-  }
-}
