--- frontend_second/.prettierrc.json
@@ -1,8 +1,0 @@
-{
-  "trailingComma": "all",
-  "tabWidth": 2,
-  "semi": false,
-  "singleQuote": true,
-  "jsxSingleQuote": true,
-  "printWidth": 100
-}
