--- frontend_second/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "fronted_second",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev -p 4000",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint --dir src",
-    "format": "prettier --write --ignore-path .gitignore './**/*.{js,jsx,ts,tsx,json}'",
-    "lint:style": "stylelint '**/*.{css,scss,sass}'",
-    "lint:style:fix": "stylelint --fix '**/*.{css,scss,sass}'"
-  },
-  "dependencies": {
-    "@types/node": "18.15.3",
-    "@types/react": "18.0.28",
-    "@types/react-dom": "18.0.11",
-    "axios": "^1.3.4",
-    "eslint": "8.36.0",
-    "eslint-config-next": "13.2.4",
-    "next": "13.2.4",
-    "next-auth": "^4.20.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "5.0.2"
-  },
-  "devDependencies": {
-    "autoprefixer": "^10.4.14",
-    "eslint-config-prettier": "^8.7.0",
-    "postcss": "^8.4.21",
-    "prettier": "^2.8.4",
-    "stylelint": "^15.3.0",
-    "stylelint-config-standard": "^31.0.0",
-    "stylelint-order": "^6.0.3",
-    "tailwindcss": "^3.2.7"
-  }
-}
