--- frontend_second/src/pages/_app.tsx
@@ -1,9 +1,0 @@
-import { SessionProvider } from 'next-auth/react'
-
-export default function App({ Component, pageProps: { session, ...pageProps } }) {
-  return (
-    <SessionProvider session={session}>
-      <Component {...pageProps} />
-    </SessionProvider>
-  )
-}
