--- frontend_second/src/pages/index.tsx
@@ -1,18 +1,0 @@
-import Head from 'next/head'
-
-export default function Home() {
-  return (
-    <>
-      <Head>
-        <title>アウトドア般若心経 | Top</title>
-      </Head>
-      <div>
-        <h1>アウトドア般若心経</h1>
-        <p>自分探しならぬ、自分なくしの旅へ</p>
-      </div>
-      <div>
-        <a href='/welcome'>welcomeページへ</a>
-      </div>
-    </>
-  )
-}
