--- frontend_second/src/styles/Home.module.css
@@ -1,280 +1,0 @@
-.main {
-  align-items: center;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 100vh;
-  padding: 6rem;
-}
-
-.description {
-  align-items: inherit;
-  display: inherit;
-  font-family: var(--font-mono);
-  font-size: 0.85rem;
-  justify-content: inherit;
-  max-width: var(--max-width);
-  width: 100%;
-  z-index: 2;
-}
-
-.description a {
-  align-items: center;
-  display: flex;
-  gap: 0.5rem;
-  justify-content: center;
-}
-
-.description p {
-  background-color: rgb(var(--callout-rgb) 0.5);
-  border: 1px solid rgb(var(--callout-border-rgb) 0.3);
-  border-radius: var(--border-radius);
-  margin: 0;
-  padding: 1rem;
-  position: relative;
-}
-
-.code {
-  font-family: var(--font-mono);
-  font-weight: 700;
-}
-
-.grid {
-  display: grid;
-  grid-template-columns: repeat(4, minmax(25%, auto));
-  max-width: 100%;
-  width: var(--max-width);
-}
-
-.card {
-  background: rgb(var(--card-rgb) 0);
-  border: 1px solid rgb(var(--card-border-rgb) 0);
-  border-radius: var(--border-radius);
-  padding: 1rem 1.2rem;
-  transition: background 200ms, border 200ms;
-}
-
-.card span {
-  display: inline-block;
-  transition: transform 200ms;
-}
-
-.card h2 {
-  font-weight: 600;
-  margin-bottom: 0.7rem;
-}
-
-.card p {
-  font-size: 0.9rem;
-  line-height: 1.5;
-  margin: 0;
-  max-width: 30ch;
-  opacity: 0.6;
-}
-
-.center {
-  align-items: center;
-  display: flex;
-  justify-content: center;
-  padding: 4rem 0;
-  position: relative;
-}
-
-.center::before {
-  background: var(--secondary-glow);
-  border-radius: 50%;
-  height: 360px;
-  margin-left: -400px;
-  width: 480px;
-}
-
-.center::after {
-  background: var(--primary-glow);
-  height: 180px;
-  width: 240px;
-  z-index: -1;
-}
-
-.center::before,
-.center::after {
-  content: '';
-  filter: blur(45px);
-  left: 50%;
-  position: absolute;
-  transform: translateZ(0);
-}
-
-.logo,
-.thirteen {
-  position: relative;
-}
-
-.thirteen {
-  align-items: center;
-  border-radius: var(--border-radius);
-  box-shadow: 0 2px 8px -1px #0000001a;
-  display: flex;
-  height: 75px;
-  justify-content: center;
-  margin-left: 16px;
-  overflow: hidden;
-  padding: 25px 10px;
-  transform: translateZ(0);
-  width: 75px;
-}
-
-.thirteen::before,
-.thirteen::after {
-  content: '';
-  position: absolute;
-  z-index: -1;
-}
-
-/* Conic Gradient Animation */
-.thirteen::before {
-  animation: 6s rotate linear infinite;
-  background: var(--tile-border);
-  height: 200%;
-  width: 200%;
-}
-
-/* Inner Square */
-.thirteen::after {
-  background: linear-gradient(
-    to bottom right,
-    rgb(var(--tile-start-rgb) 1),
-    rgb(var(--tile-end-rgb) 1)
-  );
-  background-clip: content-box;
-  border-radius: var(--border-radius);
-  inset: 0;
-  padding: 1px;
-}
-
-/* Enable hover only on non-touch devices */
-@media (hover: hover) and (pointer: fine) {
-  .card:hover {
-    background: rgb(var(--card-rgb) 0.1);
-    border: 1px solid rgb(var(--card-border-rgb) 0.15);
-  }
-
-  .card:hover span {
-    transform: translateX(4px);
-  }
-}
-
-@media (prefers-reduced-motion) {
-  .thirteen::before {
-    animation: none;
-  }
-
-  .card:hover span {
-    transform: none;
-  }
-}
-
-/* Mobile */
-@media (max-width: 700px) {
-  .content {
-    padding: 4rem;
-  }
-
-  .grid {
-    grid-template-columns: 1fr;
-    margin-bottom: 120px;
-    max-width: 320px;
-    text-align: center;
-  }
-
-  .card {
-    padding: 1rem 2.5rem;
-  }
-
-  .card h2 {
-    margin-bottom: 0.5rem;
-  }
-
-  .center {
-    padding: 8rem 0 6rem;
-  }
-
-  .center::before {
-    height: 300px;
-    transform: none;
-  }
-
-  .description {
-    font-size: 0.8rem;
-  }
-
-  .description a {
-    padding: 1rem;
-  }
-
-  .description p,
-  .description div {
-    display: flex;
-    justify-content: center;
-    position: fixed;
-    width: 100%;
-  }
-
-  .description p {
-    align-items: center;
-    backdrop-filter: blur(24px);
-    background: linear-gradient(
-      to bottom,
-      rgb(var(--background-start-rgb) 1),
-      rgb(var(--callout-rgb) 0.5)
-    );
-    background-clip: padding-box;
-    border: none;
-    border-bottom: 1px solid rgb(var(--callout-border-rgb) 0.25);
-    border-radius: 0;
-    inset: 0 0 auto;
-    padding: 2rem 1rem 1.4rem;
-  }
-
-  .description div {
-    align-items: flex-end;
-    background: linear-gradient(
-      to bottom,
-      transparent 0%,
-      rgb(var(--background-end-rgb)) 40%
-    );
-    height: 200px;
-    inset: auto 0 0;
-    padding: 2rem;
-    pointer-events: none;
-    z-index: 1;
-  }
-}
-
-/* Tablet and Smaller Desktop */
-@media (min-width: 701px) and (max-width: 1120px) {
-  .grid {
-    grid-template-columns: repeat(2, 50%);
-  }
-}
-
-@media (prefers-color-scheme: dark) {
-  /* stylelint-disable selector-class-pattern */
-  .vercelLogo {
-    filter: invert(1);
-  }
-
-  .logo,
-  .thirteen img {
-    filter: invert(1) drop-shadow(0 0 0.3rem #ffffff70);
-  }
-}
-
-@keyframes rotate {
-  from {
-    transform: rotate(360deg);
-  }
-
-  to {
-    transform: rotate(0deg);
-  }
-}
